scrolling
I’ve made a few tweaks here and there working on getting the status to
be displayed at the top as well as getting the symbols from the number
keys when shifted. The big thing is scrolling, so technically it works,
but its pretty about half the time and I have absolutely no idea why,
so I’m working on that. I also changed the OS version to “over 9000”
because, well, I can